Publish DEPS for Chromium 5.0.375.127
Merge 55319 - Added defense in depth mechanism to host resolver to prevent long hostnames from getting into glibc's getaddrinfo.

BUG=50383
TEST=HostResolverImplTest.LongHost
Review URL: http://codereview.chromium.org/3052044

Review URL: http://codereview.chromium.org/3130013

git-svn-id: svn://svn.chromium.org/chrome/branches/375/src@55887 0039d316-1c4b-4281-b951-d872f2087c98

+++ chrome_frame/sync_msg_reply_dispatcher.h
@@ -7,6 +7,7 @@
 
 #include <deque>
 
+#include "base/callback.h"
 #include "base/lock.h"
 #include "ipc/ipc_channel_proxy.h"
 
@@ -51,12 +52,12 @@ class SyncMessageReplyDispatcher : public IPC::ChannelProxy::MessageFilter {
  protected:
   struct MessageSent {
     MessageSent() {}
-    MessageSent(int id, uint16 type, void* callback, void* key)
+    MessageSent(int id, uint32 type, void* callback, void* key)
       : id(id), callback(callback), type(type), key(key) {}
     int id;
     void* callback;
     void* key;
-    uint16 type;
+    uint32 type;
   };
 
   typedef std::deque<MessageSent> PendingSyncMessageQueue;
