+++ webkit/glue/plugins/plugin_stream_url.cc
@@ -1,18 +1,17 @@
-// Copyright (c) 2006-2008 The Chromium Authors. All rights reserved.
+// Copyright (c) 2010 The Chromium Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
 #include "webkit/glue/plugins/plugin_stream_url.h"
 
-#include "webkit/glue/glue_util.h"
-#include "webkit/glue/webplugin.h"
 #include "webkit/glue/plugins/plugin_host.h"
 #include "webkit/glue/plugins/plugin_instance.h"
+#include "webkit/glue/plugins/webplugin.h"
 
 namespace NPAPI {
 
 PluginStreamUrl::PluginStreamUrl(
-    int resource_id,
+    unsigned long resource_id,
     const GURL &url,
     PluginInstance *instance,
     bool notify_needed,
